Fix Build error
Building on Pop_os resulted in an At least one entry of URL is a path (invalid in a list) warning. This seems to be caused by a forgotten newline in this function. I'm no expert at this, so please correct me if I'm wrong, but this change got my build working again.