Abort if tf->model is null in createOrtSession
If an user have migrated from an older version of this plugin, old model files may be remained and tf->modelSelection is set not to be prefixed with models.
This change abort the createOrtSession if tf->model is null and it prevent the segmentation fault of executing tf->model->populateInputOutputNames() when tf->model is null.